requirements.txt
pyTelegramBotAPI
google-ai-generativelanguage
python-dotenv
pandas
